Moving image handling away from tar handling
