+++ .travis.yml
@@ -7,7 +7,7 @@ python:
 language: go
 go:
         - 1.8.1
-go_import_path: runtimes-common
+go_import_path: github.com/GoogleCloudPlatform/runtimes-common
 
 jdk:
   - oraclejdk8
