+++ iDiff/differs/package.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/GoogleCloudPlatform/runtimes-common/iDiff/utils"
 	"github.com/golang/glog"
-	"github.com/runtimes-common/iDiff/utils"
 )
 
 //  Diffs two packages and compares their contents
