+++ iDiff/utils/tar_utils.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"archive/tar"
-	"context"
 	"encoding/json"
 	"io"
 	"io/ioutil"
@@ -10,73 +9,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/docker/docker/client"
-	"github.com/docker/docker/pkg/system"
 	"github.com/golang/glog"
 )
-
-// ImageToDir converts an image to an unpacked tar and creates a representation of that directory.
-func ImageToDir(img string) (string, string, error) {
-	cli, err := client.NewEnvClient()
-	if err != nil {
-		return "", "", err
-	}
-	tarPath, err := ImageToTar(cli, img)
-	if err != nil {
-		return "", "", err
-	}
-	err = ExtractTar(tarPath)
-	if err != nil {
-		return "", "", err
-	}
-	os.Remove(tarPath)
-	path := strings.TrimSuffix(tarPath, filepath.Ext(tarPath))
-	jsonPath := path + ".json"
-	err = DirToJSON(path, jsonPath)
-	if err != nil {
-		return "", "", err
-	}
-	return jsonPath, path, nil
-
-}
-
-// copyToFile writes the content of the reader to the specified file
-func copyToFile(outfile string, r io.Reader) error {
-	// We use sequential file access here to avoid depleting the standby list
-	// on Windows. On Linux, this is a call directly to ioutil.TempFile
-	tmpFile, err := system.TempFileSequential(filepath.Dir(outfile), ".docker_temp_")
-	if err != nil {
-		return err
-	}
-
-	tmpPath := tmpFile.Name()
-
-	_, err = io.Copy(tmpFile, r)
-	tmpFile.Close()
-
-	if err != nil {
-		os.Remove(tmpPath)
-		return err
-	}
-
-	if err = os.Rename(tmpPath, outfile); err != nil {
-		os.Remove(tmpPath)
-		return err
-	}
-
-	return nil
-}
-
-// ImageToTar writes an image to a .tar file
-func ImageToTar(cli client.APIClient, image string) (string, error) {
-	imgBytes, err := cli.ImageSave(context.Background(), []string{image})
-	if err != nil {
-		return "", err
-	}
-	defer imgBytes.Close()
-	newpath := image + ".tar"
-	return newpath, copyToFile(newpath, imgBytes)
-}
 
 // Directory stores a representaiton of a file directory.
 type Directory struct {
@@ -89,7 +23,6 @@ type Directory struct {
 func UnTar(filename string, path string) error {
 	if _, ok := os.Stat(path); ok != nil {
 		os.MkdirAll(path, 0777)
-
 	}
 
 	file, err := os.Open(filename)
@@ -167,7 +100,7 @@ func ExtractTar(path string) error {
 	return filepath.Walk(path, untarWalkFn)
 }
 
-// DirToJSON records the directory structure starting at the provided path as in a json file.
+// DirToJSON records the directory structure starting at the provided path as a json file.
 func DirToJSON(path string, target string) error {
 	var directory Directory
 	directory.Root = path
